+++ finance_app/lib/pages/goalpage.dart
@@ -1,0 +1,131 @@
+import 'package:flutter/material.dart';
+
+class GoalsScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        ListTile(
+          title: Text('Finance Goals'),
+          trailing: IconButton(
+            icon: Icon(Icons.add),
+            onPressed: () {
+              _showAddGoalDialog(context);
+            },
+          ),
+        ),
+        Expanded(
+          child: ListView.builder(
+            itemCount: 10, // replace with your goals count
+            itemBuilder: (context, index) {
+              return GoalCard();
+            },
+          ),
+        ),
+      ],
+    );
+  }
+
+  void _showAddGoalDialog(BuildContext context) {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text('Add New Goal'),
+          content: AddGoalForm(),
+          actions: <Widget>[
+            TextButton(
+              child: Text('Cancel'),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+            TextButton(
+              child: Text('Add'),
+              onPressed: () {
+                // Handle adding new goal
+                Navigator.of(context).pop();
+              },
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
+
+class GoalCard extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      margin: EdgeInsets.all(10.0),
+      child: ListTile(
+        title: Text('Goal Name'),
+        subtitle: Text('Total Amount: \$1000'),
+        trailing: TextButton(
+          child: Text('Pay'),
+          onPressed: () {
+            // Handle pay towards goal
+          },
+        ),
+      ),
+    );
+  }
+}
+
+class AddGoalForm extends StatefulWidget {
+  @override
+  _AddGoalFormState createState() => _AddGoalFormState();
+}
+
+class _AddGoalFormState extends State<AddGoalForm> {
+  final _formKey = GlobalKey<FormState>();
+  String _goalName = '';
+  String _goalType = 'Saving';
+  String _description = '';
+
+  @override
+  Widget build(BuildContext context) {
+    return Form(
+      key: _formKey,
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          TextFormField(
+            decoration: InputDecoration(labelText: 'Goal Name'),
+            validator: (value) {
+              if (value!.isEmpty) {
+                return 'Please enter a goal name';
+              }
+              return null;
+            },
+            onSaved: (value) {
+              _goalName = value!;
+            },
+          ),
+          DropdownButtonFormField<String>(
+            value: _goalType,
+            items: <String>['Saving', 'Investment', 'Miscellaneous'].map((String value) {
+              return DropdownMenuItem<String>(
+                value: value,
+                child: Text(value),
+              );
+            }).toList(),
+            decoration: InputDecoration(labelText: 'Goal Type'),
+            onChanged: (value) {
+              setState(() {
+                _goalType = value!;
+              });
+            },
+          ),
+          TextFormField(
+            decoration: InputDecoration(labelText: 'Description'),
+            onSaved: (value) {
+              _description = value!;
+            },
+          ),
+        ],
+      ),
+    );
+  }
+}
