+++ finance_app/lib/widget/section_card.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class SectionCard extends StatelessWidget {
+  final String title;
+
+  SectionCard({required this.title});
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      margin: EdgeInsets.all(10.0),
+      child: ListTile(
+        title: Text(title),
+      ),
+    );
+  }
+}
